backend/requirements.txt
# Dasko backend — Gemini Live API proxy
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
google-genai>=1.65.0
python-dotenv>=1.0.0
